fix(n11): attribute priority xs:double olarak parse — int parse "1.0"da null dönüyordu
N11 CategoryService WSDL doğrulaması (canlı): priority = xs:double. Dünkü int parse "1.0" gibi
değerlerde başarısız olup priority'yi null bırakıyordu → sıralama etkisiz kalırdı.

- N11AttributeDef.Priority + N11CategoryAttributeDto.Priority: int? → double?
- REST parse: GetInt32 → GetDouble; SOAP fallback: int.TryParse → double.TryParse (InvariantCulture).
- EditFields sıralama: ThenBy(Priority ?? double.MaxValue).

Not (WSDL bulgusu): özet method GetCategoryAttributesId (CategoryProductAttributeData) HAFİF ama
priority + valueList İÇERMEZ → edit formu için tam GetCategoryAttributes doğru seçim (değiştirilmedi).

+++ src/Integration.TradeXpress.Application.Contracts/N11Categories/N11CategoryDtos.cs
@@ -22,8 +22,8 @@ public class N11CategoryAttributeDto
     public bool IsVariant { get; set; }
     public bool IsCustomValue { get; set; }
 
-    /// <summary>N11'in form öncelik sırası (REST'te dolu; SOAP fallback'te null). UI sıralaması bunu kullanır.</summary>
-    public int? Priority { get; set; }
+    /// <summary>N11'in form öncelik sırası (WSDL: xs:double). UI sıralaması bunu kullanır; çözülemezse null → sona.</summary>
+    public double? Priority { get; set; }
 
     public List<N11CategoryAttributeValueDto> Values { get; set; } = new();
 }
